MODIFICAR SIN JOPTION

+++ Lab6_LuisFlores/src/Carrera.java
@@ -50,7 +50,7 @@ public class Carrera {
         return costo;
     }
 
-    public void setCosto(float costo) {
+    public void setCosto(double costo) {
         this.costo = costo;
     }
